WIP on Stockfish integration
Modified gameSceneController to work with Stockfish as an opponent for both white player and black player

Encountering issue where sometimes when playing as black, either stockfish or player's moves will cause an invalid pointer exception - indicating that some square that
has no children is being accessed with .get(0) - However this should never happen as Stockfish should know board state, and the User interacts through GUI by clicking a square
that does in fact contain a piece.

No predicatbility for when the error occurs, just that it is when the player plays as black
	- Likely something to do with converting the square labels to subjective square index (depending on player perspective)

Committing now before I attempt to fix this bug and break everything.