Added Stockfish move promotion parsing.
Changed UciParser.BEST_MOVE from a pattern compiled on "bestmove ([a-h][0-9][a-h][0-9])" to one compiled on "bestmove ([a-h][0-9][a-h][0-9][nbrq]?)" in order to extract the entire move with promotion.
	- UciParser.convertMove() then takes the either 4 or 5 length string and outputs either an int[2] or int[3] depending on if its a promotion. First two elements are orign, destination, third optional element is promoted piece
		- Promoted piece int is Absolute, the owner of the promoted piece is not taken to account as UciParser is static.
		- The sign is changed accordingly in GameSceneController.playOpponentMove()

Testing reading promotions from Stockfish is a bit tricky as I need to coerce Stockfish into promoting by playing in a way that guides it towards promotion.
	 - There are additional methods of coercion but that would require specific setup using 'go searchmoves ....'